+++ handler/openid/validator_test.go
@@ -225,7 +225,7 @@ func TestValidatePrompt(t *testing.T) {
 	} {
 		t.Run(fmt.Sprintf("case=%d/description=%s", k, tc.d), func(t *testing.T) {
 			t.Logf("%s", tc.idTokenHint)
-			err := v.ValidatePrompt(&fosite.AuthorizeRequest{
+			err := v.ValidatePrompt(context.TODO(), &fosite.AuthorizeRequest{
 				Request: fosite.Request{
 					Form:    url.Values{"prompt": {tc.prompt}, "id_token_hint": {tc.idTokenHint}},
 					Client:  &fosite.DefaultClient{Public: tc.isPublic},